Move tools to special directory
