+++ czkawka_cli/src/main.rs
@@ -9,22 +9,22 @@ use std::thread;
 use clap::Parser;
 use commands::Commands;
 use crossbeam_channel::{Receiver, Sender, unbounded};
-use czkawka_core::bad_extensions::{BadExtensions, BadExtensionsParameters};
-use czkawka_core::big_file::{BigFile, BigFileParameters, SearchMode};
-use czkawka_core::broken_files::{BrokenFiles, BrokenFilesParameters, CheckedTypes};
 use czkawka_core::common::{DEFAULT_THREAD_SIZE, print_version_mode, set_number_of_threads, setup_logger};
 use czkawka_core::common_tool::{CommonData, DeleteMethod};
 #[allow(unused_imports)] // It is used in release for print_results_to_output().
 use czkawka_core::common_traits::*;
-use czkawka_core::duplicate::{DuplicateFinder, DuplicateFinderParameters};
-use czkawka_core::empty_files::EmptyFiles;
-use czkawka_core::empty_folder::EmptyFolder;
-use czkawka_core::invalid_symlinks::InvalidSymlinks;
 use czkawka_core::progress_data::ProgressData;
-use czkawka_core::same_music::{SameMusic, SameMusicParameters};
-use czkawka_core::similar_images::{SimilarImages, SimilarImagesParameters, return_similarity_from_similarity_preset, test_image_conversion_speed};
-use czkawka_core::similar_videos::{SimilarVideos, SimilarVideosParameters};
-use czkawka_core::temporary::Temporary;
+use czkawka_core::tools::bad_extensions::{BadExtensions, BadExtensionsParameters};
+use czkawka_core::tools::big_file::{BigFile, BigFileParameters, SearchMode};
+use czkawka_core::tools::broken_files::{BrokenFiles, BrokenFilesParameters, CheckedTypes};
+use czkawka_core::tools::duplicate::{DuplicateFinder, DuplicateFinderParameters};
+use czkawka_core::tools::empty_files::EmptyFiles;
+use czkawka_core::tools::empty_folder::EmptyFolder;
+use czkawka_core::tools::invalid_symlinks::InvalidSymlinks;
+use czkawka_core::tools::same_music::{SameMusic, SameMusicParameters};
+use czkawka_core::tools::similar_images::{SimilarImages, SimilarImagesParameters, return_similarity_from_similarity_preset, test_image_conversion_speed};
+use czkawka_core::tools::similar_videos::{SimilarVideos, SimilarVideosParameters};
+use czkawka_core::tools::temporary::Temporary;
 use log::error;
 
 use crate::commands::{
