+++ czkawka_core/src/tools/bad_extensions.rs
