+++ czkawka_core/src/tools/big_file.rs
