+++ czkawka_core/src/tools/broken_files.rs
