+++ czkawka_core/src/tools/empty_files.rs
