+++ czkawka_core/src/tools/empty_folder.rs
