+++ czkawka_core/src/tools/invalid_symlinks.rs
