+++ czkawka_core/src/tools/mod.rs
@@ -1,0 +1,11 @@
+pub mod bad_extensions;
+pub mod big_file;
+pub mod broken_files;
+pub mod duplicate;
+pub mod empty_files;
+pub mod empty_folder;
+pub mod invalid_symlinks;
+pub mod same_music;
+pub mod similar_images;
+pub mod similar_videos;
+pub mod temporary;
