+++ czkawka_core/src/tools/similar_videos.rs
