+++ czkawka_core/src/tools/temporary.rs
