+++ czkawka_gui/src/connect_things/connect_button_search.rs
@@ -4,21 +4,21 @@ use std::sync::atomic::{AtomicBool, Ordering};
 use std::thread;
 
 use crossbeam_channel::Sender;
-use czkawka_core::bad_extensions::{BadExtensions, BadExtensionsParameters};
-use czkawka_core::big_file::{BigFile, BigFileParameters};
-use czkawka_core::broken_files::{BrokenFiles, BrokenFilesParameters, CheckedTypes};
 use czkawka_core::common::DEFAULT_THREAD_SIZE;
 use czkawka_core::common_dir_traversal::CheckingMethod;
 use czkawka_core::common_tool::CommonData;
-use czkawka_core::duplicate::{DuplicateFinder, DuplicateFinderParameters};
-use czkawka_core::empty_files::EmptyFiles;
-use czkawka_core::empty_folder::EmptyFolder;
-use czkawka_core::invalid_symlinks::InvalidSymlinks;
 use czkawka_core::progress_data::ProgressData;
-use czkawka_core::same_music::{MusicSimilarity, SameMusic, SameMusicParameters};
-use czkawka_core::similar_images::{SimilarImages, SimilarImagesParameters};
-use czkawka_core::similar_videos::{SimilarVideos, SimilarVideosParameters};
-use czkawka_core::temporary::Temporary;
+use czkawka_core::tools::bad_extensions::{BadExtensions, BadExtensionsParameters};
+use czkawka_core::tools::big_file::{BigFile, BigFileParameters};
+use czkawka_core::tools::broken_files::{BrokenFiles, BrokenFilesParameters, CheckedTypes};
+use czkawka_core::tools::duplicate::{DuplicateFinder, DuplicateFinderParameters};
+use czkawka_core::tools::empty_files::EmptyFiles;
+use czkawka_core::tools::empty_folder::EmptyFolder;
+use czkawka_core::tools::invalid_symlinks::InvalidSymlinks;
+use czkawka_core::tools::same_music::{MusicSimilarity, SameMusic, SameMusicParameters};
+use czkawka_core::tools::similar_images::{SimilarImages, SimilarImagesParameters};
+use czkawka_core::tools::similar_videos::{SimilarVideos, SimilarVideosParameters};
+use czkawka_core::tools::temporary::Temporary;
 use fun_time::fun_time;
 use gtk4::Grid;
 use gtk4::prelude::*;
