+++ czkawka_gui/src/connect_things/connect_similar_image_size_change.rs
@@ -1,4 +1,4 @@
-use czkawka_core::similar_images::{SIMILAR_VALUES, get_string_from_similarity};
+use czkawka_core::tools::similar_images::{SIMILAR_VALUES, get_string_from_similarity};
 use gtk4::prelude::*;
 
 use crate::gui_structs::gui_data::GuiData;
