+++ czkawka_gui/src/gui_structs/gui_main_notebook.rs
@@ -1,7 +1,7 @@
-use czkawka_core::big_file::SearchMode;
 use czkawka_core::common_dir_traversal::CheckingMethod;
 use czkawka_core::localizer_core::{fnc_get_similarity_minimal, fnc_get_similarity_very_high};
-use czkawka_core::similar_images::{SIMILAR_VALUES, get_string_from_similarity};
+use czkawka_core::tools::big_file::SearchMode;
+use czkawka_core::tools::similar_images::{SIMILAR_VALUES, get_string_from_similarity};
 use gtk4::prelude::*;
 use gtk4::{Builder, CheckButton, ComboBoxText, Entry, EventControllerKey, GestureClick, Image, Label, Notebook, Scale, ScrolledWindow, TreeView, Widget};
