+++ czkawka_gui/src/help_combo_box.rs
@@ -1,6 +1,6 @@
-use czkawka_core::big_file::SearchMode;
 use czkawka_core::common_dir_traversal::CheckingMethod;
-use czkawka_core::duplicate::HashType;
+use czkawka_core::tools::big_file::SearchMode;
+use czkawka_core::tools::duplicate::HashType;
 use image_hasher::{FilterType, HashAlg};
 
 pub struct HashTypeStruct {
