+++ czkawka_gui/src/initialize_gui.rs
@@ -2,8 +2,8 @@ use std::cell::RefCell;
 use std::rc::Rc;
 
 use czkawka_core::common_image::get_dynamic_image_from_path;
-use czkawka_core::similar_images::SIMILAR_VALUES;
-use czkawka_core::similar_videos::MAX_TOLERANCE;
+use czkawka_core::tools::similar_images::SIMILAR_VALUES;
+use czkawka_core::tools::similar_videos::MAX_TOLERANCE;
 use gdk4::gdk_pixbuf::Pixbuf;
 use glib::types::Type;
 use gtk4::gdk_pixbuf::InterpType;
