+++ czkawka_gui/src/saving_loading.rs
@@ -8,7 +8,7 @@ use std::{env, fs};
 use czkawka_core::common::get_all_available_threads;
 use czkawka_core::common_dir_traversal::CheckingMethod;
 use czkawka_core::common_items::DEFAULT_EXCLUDED_ITEMS;
-use czkawka_core::similar_images::SIMILAR_VALUES;
+use czkawka_core::tools::similar_images::SIMILAR_VALUES;
 use directories_next::ProjectDirs;
 use gtk4::prelude::*;
 use gtk4::{ComboBoxText, ScrolledWindow, TextView, TreeView};
