+++ krokiet/src/connect_scan.rs
@@ -5,23 +5,23 @@ use std::thread;
 
 use chrono::DateTime;
 use crossbeam_channel::Sender;
-use czkawka_core::bad_extensions::{BadExtensions, BadExtensionsParameters, BadFileEntry};
-use czkawka_core::big_file::{BigFile, BigFileParameters, SearchMode};
-use czkawka_core::broken_files::{BrokenEntry, BrokenFiles, BrokenFilesParameters, CheckedTypes};
 use czkawka_core::common::{DEFAULT_THREAD_SIZE, split_path, split_path_compare};
 use czkawka_core::common_dir_traversal::{CheckingMethod, FileEntry};
 use czkawka_core::common_tool::CommonData;
 use czkawka_core::common_traits::ResultEntry;
-use czkawka_core::duplicate::{DuplicateEntry, DuplicateFinder, DuplicateFinderParameters};
-use czkawka_core::empty_files::EmptyFiles;
-use czkawka_core::empty_folder::{EmptyFolder, FolderEntry};
-use czkawka_core::invalid_symlinks::{InvalidSymlinks, SymlinksFileEntry};
 use czkawka_core::progress_data::ProgressData;
-use czkawka_core::same_music::{MusicEntry, MusicSimilarity, SameMusic, SameMusicParameters};
-use czkawka_core::similar_images;
-use czkawka_core::similar_images::{ImagesEntry, SimilarImages, SimilarImagesParameters};
-use czkawka_core::similar_videos::{SimilarVideos, SimilarVideosParameters, VideosEntry};
-use czkawka_core::temporary::{Temporary, TemporaryFileEntry};
+use czkawka_core::tools::bad_extensions::{BadExtensions, BadExtensionsParameters, BadFileEntry};
+use czkawka_core::tools::big_file::{BigFile, BigFileParameters, SearchMode};
+use czkawka_core::tools::broken_files::{BrokenEntry, BrokenFiles, BrokenFilesParameters, CheckedTypes};
+use czkawka_core::tools::duplicate::{DuplicateEntry, DuplicateFinder, DuplicateFinderParameters};
+use czkawka_core::tools::empty_files::EmptyFiles;
+use czkawka_core::tools::empty_folder::{EmptyFolder, FolderEntry};
+use czkawka_core::tools::invalid_symlinks::{InvalidSymlinks, SymlinksFileEntry};
+use czkawka_core::tools::same_music::{MusicEntry, MusicSimilarity, SameMusic, SameMusicParameters};
+use czkawka_core::tools::similar_images;
+use czkawka_core::tools::similar_images::{ImagesEntry, SimilarImages, SimilarImagesParameters};
+use czkawka_core::tools::similar_videos::{SimilarVideos, SimilarVideosParameters, VideosEntry};
+use czkawka_core::tools::temporary::{Temporary, TemporaryFileEntry};
 use humansize::{BINARY, format_size};
 use rayon::prelude::*;
 use slint::{ComponentHandle, ModelRc, SharedString, VecModel, Weak};
